+++ CHANGELOG.md
@@ -23,6 +23,16 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0).
 
 <!-- Types of changes: Added, Changed, Deprecated, Removed, Fixed -->
 
+## [0.8.0] - 03/09/2018
+### Added:
+- Legal move indicators
+- Mouseover coordinates
+- Check for absolute pins
+- About PyQt
+### Fixed:
+- Bug allowing castling when in check
+
+
 ## [0.7.0] - 24/08/2018
 ### Added:
 - Promote variation (right-click context menu)
@@ -98,6 +108,7 @@ Initial development release.
 
 
 <!-- Links to releases -->
+[0.8.0]: https://github.com/GammaDeltaII/4PlayerChess/releases/tag/0.8.0
 [0.7.0]: https://github.com/GammaDeltaII/4PlayerChess/releases/tag/0.7.0
 [0.6.1]: https://github.com/GammaDeltaII/4PlayerChess/releases/tag/0.6.1
 [0.6.0]: https://github.com/GammaDeltaII/4PlayerChess/releases/tag/0.6.0